requirements.txt
django
channels
djangorestframework
djangorestframework-simplejwt
psycopg2
Pillow
